Updated code

+++ src/main/kotlin/com/example/AgroNearBackend/Repository/UserRepository.kt
@@ -3,9 +3,7 @@ package com.example.AgroNearBackend.repository
 import com.example.AgroNearBackend.Entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository : JpaRepository<User, Int> {
-
+interface UserRepository : JpaRepository<User, Long> {
     fun findByEmail(email: String): User?
-
     fun existsByEmail(email: String): Boolean
 }
